Support for loading animations
This update delays installing the canvas until after the
images are available. It also removes all of the host element's
existing children before installing the canvas.

This allows the host page to add a "loading" animation as a
child of the host element, which then gets removed when the
loading is finished.